Rename `Observer` to `ClientObserver`.
